Move AST helpers out of main
Move AST helpers out of main and into ASTUtils.
