+++ lib/MakiASTConsumer.cc
@@ -34,7 +34,6 @@
 #include <functional>
 #include <llvm/ADT/StringRef.h>
 #include <llvm/Support/Casting.h>
-#include <queue>
 #include <set>
 #include <string>
 #include <tuple>
@@ -43,235 +42,6 @@
 
 namespace maki {
 using namespace clang::ast_matchers;
-
-// Collect all subtrees of the given stmt using BFS
-std::set<const clang::Stmt *> subtrees(const clang::Stmt *ST) {
-    std::set<const clang::Stmt *> Subtrees;
-    if (!ST) {
-        return Subtrees;
-    }
-
-    std::queue<const clang::Stmt *> Q({ ST });
-    while (!Q.empty()) {
-        auto Cur = Q.front();
-        Q.pop();
-        if (Cur) {
-            Subtrees.insert(Cur);
-            for (auto &&Child : Cur->children()) {
-                Q.push(Child);
-            }
-        }
-    }
-    return Subtrees;
-}
-
-// Returns true if LHS is a subtree of RHS via BFS
-bool inTree(const clang::Stmt *LHS, const clang::Stmt *RHS) {
-    std::queue<const clang::Stmt *> Q({ RHS });
-    while (!Q.empty()) {
-        auto Cur = Q.front();
-        Q.pop();
-        if (LHS == Cur) {
-            return true;
-        }
-        if (Cur) {
-            for (auto &&Child : Cur->children()) {
-                Q.push(Child);
-            }
-        }
-    }
-    return false;
-}
-
-// Returns true if the given predicate returns true for any type
-// contained in the given type
-bool isInType(const clang::Type *T, clang::ASTContext &Ctx,
-              std::function<bool(const clang::Type *)> pred) {
-    debug("Checking if T is a pointer type");
-    while (T && (T->isAnyPointerType() || T->isArrayType())) {
-        if (T->isAnyPointerType()) {
-            T = T->getPointeeType().getTypePtrOrNull();
-        } else if (T->isArrayType()) {
-            T = T->getBaseElementTypeUnsafe();
-        }
-    }
-
-    return pred(T);
-}
-
-clang::Decl *getTypeDeclOrNull(const clang::Type *T) {
-    if (!T) {
-        return nullptr;
-    }
-
-    if (auto TD = clang::dyn_cast<clang::TypedefType>(T)) {
-        return getTypeDeclOrNull(TD->desugar().getTypePtrOrNull());
-    } else if (auto TD = clang::dyn_cast<clang::TagType>(T)) {
-        return TD->getDecl();
-    } else if (auto ET = clang::dyn_cast<clang::ElaboratedType>(T)) {
-        return getTypeDeclOrNull(ET->desugar().getTypePtrOrNull());
-    } else {
-        return nullptr;
-    }
-}
-
-// Returns true if any type in T was defined after L
-bool hasTypeDefinedAfter(const clang::Type *T, clang::ASTContext &Ctx,
-                         clang::SourceLocation L) {
-    auto &SM = Ctx.getSourceManager();
-    return isInType(T, Ctx, [&SM, L](const clang::Type *T) {
-        if (!T) {
-            return false;
-        }
-
-        auto *D = getTypeDeclOrNull(T);
-
-        if (!D) {
-            return false;
-        }
-
-        auto DLoc = D->getLocation();
-        if (DLoc.isInvalid()) {
-            return false;
-        }
-
-        auto DFLoc = SM.getFileLoc(DLoc);
-        if (DFLoc.isInvalid()) {
-            return false;
-        }
-
-        return SM.isBeforeInTranslationUnit(L, DFLoc);
-    });
-}
-
-// Returns true if any type in T is an anonymous type
-bool hasAnonymousType(const clang::Type *T, clang::ASTContext &Ctx) {
-    return isInType(T, Ctx, [](const clang::Type *T) {
-        if (!T) {
-            return false;
-        }
-
-        auto D = getTypeDeclOrNull(T);
-        if (!D) {
-            return false;
-        }
-
-        auto ND = clang::dyn_cast<clang::NamedDecl>(D);
-        if (!ND) {
-            return false;
-        }
-
-        return ND->getIdentifier() == nullptr || ND->getName().empty();
-    });
-}
-
-// Returns true if any type in T is a local type
-bool hasLocalType(const clang::Type *T, clang::ASTContext &Ctx) {
-    return isInType(T, Ctx, [](const clang::Type *T) {
-        if (!T) {
-            return false;
-        }
-
-        auto D = getTypeDeclOrNull(T);
-        if (!D) {
-            return false;
-        }
-
-        auto DCtx = D->getDeclContext();
-
-        return !DCtx->isTranslationUnit();
-    });
-}
-
-// Returns true if ST is a descendant of a Node which can only have
-// subexpressions that are integral constants expressions
-bool isDescendantOfNodeRequiringICE(clang::ASTContext &Ctx,
-                                    const clang::Stmt *ST) {
-    if (!ST) {
-        return false;
-    }
-
-    std::queue<clang::DynTypedNode> Q;
-    for (auto P : Ctx.getParents(*ST)) {
-        Q.push(P);
-    }
-    while (!Q.empty()) {
-        auto Cur = Q.front();
-        Q.pop();
-
-        if (Cur.get<clang::CaseStmt>() || Cur.get<clang::EnumDecl>()) {
-            return true;
-        }
-
-        if (auto FD = Cur.get<clang::FieldDecl>()) {
-            if (FD->isBitField()) {
-                return true;
-            }
-            if (auto Type = FD->getType(); clang::isa<clang::ArrayType>(Type)) {
-                return true;
-            }
-        }
-
-        if (auto VD = Cur.get<clang::ValueDecl>()) {
-            // Check if this expression is the size specifier for an array type
-            // for a value declaration.
-            if (auto Type = VD->getType(); Type->isConstantArrayType()) {
-                return true;
-            } else if (auto VarD = clang::dyn_cast<clang::VarDecl>(VD)) {
-                // Check if the expression is the initializer of an integer
-                // variable that has global or static storage.
-                if (VarD->hasGlobalStorage() || VarD->isStaticLocal()) {
-                    if (VarD->getType()->isIntegerType()) {
-                        return true;
-                    }
-                }
-            }
-        }
-
-        // Check if this expression is the size specifier for an array type
-        // inside a typedef.
-        if (auto TD = Cur.get<clang::TypedefDecl>()) {
-            if (TD->getUnderlyingType()->isConstantArrayType()) {
-                return true;
-            }
-        }
-
-        for (auto P : Ctx.getParents(Cur)) {
-            Q.push(P);
-        }
-    }
-    return false;
-}
-
-// Returns true if ST is a descendant of a Node which can only have
-// subexpressions that are constant expressions.
-bool isDescendantOfNodeRequiringConstantExpression(clang::ASTContext &Ctx,
-                                                   const clang::Stmt *ST) {
-    if (!ST) {
-        return false;
-    }
-
-    std::queue<clang::DynTypedNode> Q;
-    for (auto P : Ctx.getParents(*ST)) {
-        Q.push(P);
-    }
-
-    while (!Q.empty()) {
-        auto Cur = Q.front();
-        Q.pop();
-        if (auto VD = Cur.get<clang::VarDecl>()) {
-            // Check if the expression is the initializer of a variable that has
-            // global or static storage.
-            if (VD->hasGlobalStorage() || VD->isStaticLocal()) {
-                return true;
-            }
-        }
-        for (auto P : Ctx.getParents(Cur)) {
-            Q.push(P);
-        }
-    }
-    return false;
-}
 
 // Checks if the included file is a globally included file.
 // The first element of the return result if false if not;
